+++ src/components/CardAccordion/CardAccordion.jsx
@@ -1,19 +1,19 @@
 import React, { Component } from 'react';
 import { Accordion, Card } from 'react-bootstrap';
 
-class ReportAccordion extends Component {
+class CardAccordion extends Component {
 
     constructor(props) {
         super(props);
 
         this.state = {
-            accordionOpen: true
+            accordionOpen: props.defaultOpen
         }
     }
 
     render() {
         return (
-            <Accordion defaultActiveKey={"0"} className="mb-3">
+            <Accordion defaultActiveKey={this.props.defaultOpen ? "0" : ""} className="mb-3">
                 <Card>
                     <Accordion.Toggle
                         className="report-toggle"
@@ -21,8 +21,8 @@ class ReportAccordion extends Component {
                         eventKey={"0"}
                         style={{ cursor: 'pointer' }}
                         onClick={() => this.setState({ accordionOpen: !this.state.accordionOpen })}>
-                        {this.props.searchCardHeader}
-                            <div className={this.state.accordionOpen ? 'arrow-down float-right' : 'arrow-right float-right'} />
+                        {this.props.cardHeader}
+                        <div className={this.state.accordionOpen ? 'arrow-down float-right' : 'arrow-right float-right'} />
                     </Accordion.Toggle>
                     <Accordion.Collapse eventKey="0">
                         <Card.Body>{this.props.children}</Card.Body>
@@ -33,4 +33,4 @@ class ReportAccordion extends Component {
     }
 }
 
-export default ReportAccordion;
+export default CardAccordion;